+++ test/test_stack/test_tdigest.py
@@ -1,15 +1,22 @@
 import math
+from math import inf
+
 import pytest
 import redis
-from math import inf
 
 from test.testtools import get_protocol_version
 
 topk_tests = pytest.importorskip("probables")
 
+pytestmark = []
+pytestmark.extend(
+    [
+        pytest.mark.unsupported_server_types("dragonfly"),
+    ]
+)
+
 
 @pytest.mark.min_server("7")
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_reset(r: redis.Redis):
     assert r.tdigest().create("tDigest", 10)
     # reset on empty histogram
@@ -24,7 +31,6 @@ def test_tdigest_reset(r: redis.Redis):
 
 
 @pytest.mark.min_server("7")
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_merge(r: redis.Redis):
     assert r.tdigest().create("to-tDigest", 10)
     assert r.tdigest().create("from-tDigest", 10)
@@ -50,7 +56,6 @@ def test_tdigest_merge(r: redis.Redis):
     assert r.tdigest().max("to-tDigest") == 4.0
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_min_and_max(r: redis.Redis):
     assert r.tdigest().create("tDigest", 100)
     # insert data-points into sketch
@@ -60,7 +65,6 @@ def test_tdigest_min_and_max(r: redis.Redis):
     assert 1 == r.tdigest().min("tDigest")
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_quantile(r: redis.Redis):
     assert r.tdigest().create("tDigest", 500)
     # insert data-points into sketch
@@ -80,7 +84,6 @@ def test_tdigest_quantile(r: redis.Redis):
     assert [3.0, 5.0] == r.tdigest().quantile("t-digest", 0.5, 0.8)
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_cdf(r: redis.Redis):
     assert r.tdigest().create("tDigest", 100)
     # insert data-points into sketch
@@ -91,7 +94,6 @@ def test_tdigest_cdf(r: redis.Redis):
     assert [0.1, 0.9] == [round(x, 1) for x in res]
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_trimmed_mean(r: redis.Redis):
     assert r.tdigest().create("tDigest", 100)
     # insert data-points into sketch
@@ -100,7 +102,6 @@ def test_tdigest_trimmed_mean(r: redis.Redis):
     assert 4.5 == r.tdigest().trimmed_mean("tDigest", 0.4, 0.5)
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_rank(r: redis.Redis):
     assert r.tdigest().create("t-digest", 500)
     assert r.tdigest().add("t-digest", list(range(0, 20)))
@@ -110,7 +111,6 @@ def test_tdigest_rank(r: redis.Redis):
     assert [-1, 20, 9] == r.tdigest().rank("t-digest", -20, 20, 9)
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_revrank(r: redis.Redis):
     assert r.tdigest().create("t-digest", 500)
     assert r.tdigest().add("t-digest", list(range(0, 20)))
@@ -119,7 +119,6 @@ def test_tdigest_revrank(r: redis.Redis):
     assert [-1, 19, 9] == r.tdigest().revrank("t-digest", 21, 0, 10)
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_byrank(r: redis.Redis):
     assert r.tdigest().create("t-digest", 500)
     assert r.tdigest().add("t-digest", list(range(1, 11)))
@@ -130,7 +129,6 @@ def test_tdigest_byrank(r: redis.Redis):
         r.tdigest().byrank("t-digest", -1)[0]
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_byrevrank(r: redis.Redis):
     assert r.tdigest().create("t-digest", 500)
     assert r.tdigest().add("t-digest", list(range(1, 11)))
@@ -141,7 +139,6 @@ def test_tdigest_byrevrank(r: redis.Redis):
         r.tdigest().byrevrank("t-digest", -1)[0]
 
 
-@pytest.mark.unsupported_server_types("dragonfly")
 def test_tdigest_quantile_nan(r: redis.Redis):
     r.tdigest().create("foo")
     r.tdigest().add("foo", [123])