+++ docs/about/changelog.md
@@ -7,7 +7,14 @@ tags:
 toc_depth: 2
 ---
 
-## v2.32.1 -
+## v2.32.2 -
+
+### 🐛 Bug Fixes
+
+- fix: FakeAsyncRedis supports protocol=3 #442
+- fix: TcpFakeServer supports pub/sub #431
+
+## v2.32.1 - 2025-11-05
 
 ### 🐛 Bug Fixes